pyproject.toml
[project]
name = "olhocao"
version = "1.0.0"
description = "OlhóCão Centro Canino | Website"
authors = [
    {name = "Daniel Pinto",email = "dmp593@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-browser-reload (>=1.18.0,<2.0.0)",
    "django-modeltranslation (>=0.19.12,<0.20.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "stripe (>=12.0.0,<13.0.0)",
    "deepl (>=1.21.1,<2.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
